backend/requirements.txt
aenum==3.1.17
aiobotocore==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==4.0.3
attrs==26.1.0
boto3==1.42.91
botocore==1.42.91
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
dacite==1.9.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.136.1
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.2
google-genai==1.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
jiter==0.14.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-aws==1.4.5
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-genai==4.2.1
langchain-openai==1.1.16
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langcodes==3.5.1
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.2.6
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
overloading==0.5.0
packaging==26.2
pandas==2.3.3
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyhumps==3.8.0
pymongo==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
s3transfer==0.16.1
sap-ai-sdk-base==3.4.0
sap-ai-sdk-core==3.3.0
sap-ai-sdk-gen==6.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
